requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiolimiter==1.1.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
anytree==2.12.1
asttokens==2.4.1
async-timeout==4.0.3
async-tkinter-loop==0.9.3
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.0
beartype==0.18.5
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
datashaper==0.0.49
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
environs==11.0.0
exceptiongroup==1.2.2
executing==2.1.0
fnllm==0.0.10
fonttools==4.54.1
frozenlist==1.5.0
future==1.0.0
gensim==4.3.3
graphrag==0.9.0
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.28.1
hyppo==0.4.0
idna==3.10
isodate==0.7.2
jiter==0.6.1
joblib==1.4.2
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lancedb==0.17.0
langchain==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
llvmlite==0.43.0
markdown-it-py==3.0.0
marshmallow==3.23.0
matplotlib==3.9.2
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
networkx==3.4.1
nltk==3.9.1
Nuitka==2.5.1
numba==0.60.0
numpy==1.26.4
openai==1.57.1
ordered-set==4.1.0
orjson==3.10.11
overrides==7.7.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pefile==2023.2.7
pillow==11.0.0
pipdeptree==2.24.0
portalocker==2.10.1
POT==0.9.4
propcache==0.2.0
py==1.11.0
py2exe==0.13.0.2
pyaml-env==1.2.1
pyarrow==15.0.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.9.0
pylance==0.20.0
pynndescent==0.5.13
pyparsing==3.2.0
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.4
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
SQLAlchemy==2.0.36
statsmodels==0.14.4
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tkinter-async-execute==1.3.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.1
zstandard==0.23.0
